requirements.txt
playwright
beautifulsoup4
lxml
requests
pyannote.audio
pydub
onnxruntime
torch
resemblyzer 
numpy